cautously optimistic - integration compiling, tests succeeding
